requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
requests==2.31.0
openpyxl==3.1.2 